Fix bug in old ARM documentation.
In ARMv5TE reference manual from 2000, the pseudocode for all variants
(IA, IB, DA, DB) is correct, but the description of bit P in sec. 5.4.1
(p. A5-48) is wrong (the meanings are swapped). In newer documentation
(from 2005), this is fixed. I implemented according to the big picture,
I did not read the pseudocode carefully.